Revert "Use different package name between restart/rolling bwc tests (#1096) (#1107)"
This reverts commit 173983409ba914a3e12d1423d45ef866adf723fa.
