+++ qa/rolling-upgrade/src/test/java/org/opensearch/neuralsearch/bwc/AbstractRollingUpgradeTestCase.java
@@ -2,7 +2,7 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.neuralsearch.bwc.restart;
+package org.opensearch.neuralsearch.bwc;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -12,13 +12,16 @@ import org.junit.Before;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.neuralsearch.BaseNeuralSearchIT;
 import static org.opensearch.neuralsearch.util.TestUtils.NEURAL_SEARCH_BWC_PREFIX;
-import static org.opensearch.neuralsearch.util.TestUtils.CLIENT_TIMEOUT_VALUE;
-import static org.opensearch.neuralsearch.util.TestUtils.RESTART_UPGRADE_OLD_CLUSTER;
+import static org.opensearch.neuralsearch.util.TestUtils.OLD_CLUSTER;
+import static org.opensearch.neuralsearch.util.TestUtils.MIXED_CLUSTER;
+import static org.opensearch.neuralsearch.util.TestUtils.UPGRADED_CLUSTER;
+import static org.opensearch.neuralsearch.util.TestUtils.ROLLING_UPGRADE_FIRST_ROUND;
+import static org.opensearch.neuralsearch.util.TestUtils.BWCSUITE_CLUSTER;
 import static org.opensearch.neuralsearch.util.TestUtils.BWC_VERSION;
 import static org.opensearch.neuralsearch.util.TestUtils.generateModelId;
 import org.opensearch.test.rest.OpenSearchRestTestCase;
 
-public abstract class AbstractRestartUpgradeRestTestCase extends BaseNeuralSearchIT {
+public abstract class AbstractRollingUpgradeTestCase extends BaseNeuralSearchIT {
 
     @Before
     protected String getIndexNameForTest() {
@@ -49,12 +52,35 @@ public abstract class AbstractRestartUpgradeRestTestCase extends BaseNeuralSearc
             // increase the timeout here to 90 seconds to handle long waits for a green
             // cluster health. the waits for green need to be longer than a minute to
             // account for delayed shards
-            .put(OpenSearchRestTestCase.CLIENT_SOCKET_TIMEOUT, CLIENT_TIMEOUT_VALUE)
+            .put(OpenSearchRestTestCase.CLIENT_SOCKET_TIMEOUT, "120s")
             .build();
     }
 
-    protected static final boolean isRunningAgainstOldCluster() {
-        return Boolean.parseBoolean(System.getProperty(RESTART_UPGRADE_OLD_CLUSTER));
+    protected enum ClusterType {
+        OLD,
+        MIXED,
+        UPGRADED;
+
+        public static ClusterType instance(String value) {
+            switch (value) {
+                case OLD_CLUSTER:
+                    return OLD;
+                case MIXED_CLUSTER:
+                    return MIXED;
+                case UPGRADED_CLUSTER:
+                    return UPGRADED;
+                default:
+                    throw new IllegalArgumentException("unknown cluster type: " + value);
+            }
+        }
+    }
+
+    protected final ClusterType getClusterType() {
+        return ClusterType.instance(System.getProperty(BWCSUITE_CLUSTER));
+    }
+
+    protected final boolean isFirstMixedRound() {
+        return Boolean.parseBoolean(System.getProperty(ROLLING_UPGRADE_FIRST_ROUND, "false"));
     }
 
     protected final Optional<String> getBWCVersion() {
@@ -66,7 +92,7 @@ public abstract class AbstractRestartUpgradeRestTestCase extends BaseNeuralSearc
         return registerModelGroupAndGetModelId(requestBody);
     }
 
-    protected String registerModelGroupAndGetModelId(final String requestBody) throws Exception {
+    protected String registerModelGroupAndGetModelId(String requestBody) throws Exception {
         String modelGroupRegisterRequestBody = Files.readString(
             Path.of(classLoader.getResource("processor/CreateModelGroupRequestBody.json").toURI())
         );
@@ -74,8 +100,20 @@ public abstract class AbstractRestartUpgradeRestTestCase extends BaseNeuralSearc
         return uploadModel(String.format(LOCALE, requestBody, modelGroupId));
     }
 
-    protected void createPipelineProcessor(final String modelId, final String pipelineName) throws Exception {
+    protected void createPipelineProcessor(String modelId, String pipelineName) throws Exception {
         String requestBody = Files.readString(Path.of(classLoader.getResource("processor/PipelineConfiguration.json").toURI()));
+        createPipelineProcessor(requestBody, pipelineName, modelId, null);
+    }
+
+    protected String uploadTextImageEmbeddingModel() throws Exception {
+        String requestBody = Files.readString(Path.of(classLoader.getResource("processor/UploadModelRequestBody.json").toURI()));
+        return registerModelGroupAndGetModelId(requestBody);
+    }
+
+    protected void createPipelineForTextImageProcessor(String modelId, String pipelineName) throws Exception {
+        String requestBody = Files.readString(
+            Path.of(classLoader.getResource("processor/PipelineForTextImageProcessorConfiguration.json").toURI())
+        );
         createPipelineProcessor(requestBody, pipelineName, modelId, null);
     }
 
@@ -86,23 +124,31 @@ public abstract class AbstractRestartUpgradeRestTestCase extends BaseNeuralSearc
         return registerModelGroupAndGetModelId(requestBody);
     }
 
-    protected void createPipelineForTextImageProcessor(final String modelId, final String pipelineName) throws Exception {
-        String requestBody = Files.readString(
-            Path.of(classLoader.getResource("processor/PipelineForTextImageProcessorConfiguration.json").toURI())
-        );
-        createPipelineProcessor(requestBody, pipelineName, modelId, null);
-    }
-
-    protected void createPipelineForSparseEncodingProcessor(final String modelId, final String pipelineName, final Integer batchSize)
-        throws Exception {
+    protected void createPipelineForSparseEncodingProcessor(String modelId, String pipelineName, Integer batchSize) throws Exception {
         String requestBody = Files.readString(
             Path.of(classLoader.getResource("processor/PipelineForSparseEncodingProcessorConfiguration.json").toURI())
         );
-        createPipelineProcessor(requestBody, pipelineName, modelId, batchSize);
+        final String batchSizeTag = "{{batch_size}}";
+        if (requestBody.contains(batchSizeTag)) {
+            if (batchSize != null) {
+                requestBody = requestBody.replace(batchSizeTag, String.format(LOCALE, "\n\"batch_size\": %d,\n", batchSize));
+            } else {
+                requestBody = requestBody.replace(batchSizeTag, "");
+            }
+        }
+        createPipelineProcessor(requestBody, pipelineName, modelId, null);
     }
 
-    protected void createPipelineForSparseEncodingProcessor(final String modelId, final String pipelineName) throws Exception {
+    protected void createPipelineForSparseEncodingProcessor(String modelId, String pipelineName) throws Exception {
         createPipelineForSparseEncodingProcessor(modelId, pipelineName, null);
+    }
+
+    @Override
+    protected void updateClusterSettings() {
+        updateClusterSettings("plugins.ml_commons.only_run_on_ml_node", false);
+        // default threshold for native circuit breaker is 90, it may be not enough on test runner machine
+        updateClusterSettings("plugins.ml_commons.native_memory_threshold", 100);
+        updateClusterSettings("plugins.ml_commons.allow_registering_model_via_url", true);
     }
 
     protected void createPipelineForTextChunkingProcessor(String pipelineName) throws Exception {
